FLOW-12877: move entity bin service into entity service
