+++ src/http/APIMapping.ts
@@ -101,7 +101,6 @@ const APIMapping = {
     supportService: new APIService('support-service'),
     everSignService: new APIService('eversign-service'),
     mailchimpService: new APIService('mailchimp-sync-integration-service'),
-    entityBinService: new APIService('entity-bin-service'),
 };
 
 export default APIMapping;
