--- src/service/EntityBinService/EntityBinService.Types.ts
@@ -1,27 +1,0 @@
-import { Entity } from '@flowfact/types';
-
-export interface TrashedEntity {
-    content: Entity;
-    schemaName: string;
-    entityId: string;
-    deletedAt: number;
-    deletedBy: string;
-}
-
-export interface TrashedEntitiesResponse {
-    entries: TrashedEntity[];
-    totalCount: number;
-    page: number;
-    pageSize: number;
-    offset: number;
-    size: number;
-}
-
-export interface TrashedEntityRequestItem {
-    entityId: string;
-    schema: string;
-}
-
-export interface TrashedEntitiesSchemaNameResponse {
-    schemas: string[];
-}
